fix 'object disposed' errors in the upcoming SMAPI 3.9.2
SMAPI will dispose some assets more aggressively, so caching a direct reference to an NPC's sprite texture is unsafe. This commit fixes that by loading the sprite's texture name (instead of a hardcoded asset path) so we rarely need the fallback logic, and falling back to a placeholder sprite with a warning if that fails.